+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 1.0.2 ###
- - Added taxCategories to productTypes
+ - Added taxCategories and shippingCategories to productTypes
 
 ### 1.0.1 ###
  - Also sync what emails are triggered for an order status